Add -XX:+PortableSharedCache to JAVA_TOOL_OPTIONS env var
OpenJ9 JVM needs to create "portable" AOT code when run in containers.
Normally, OpenJ9 detects that it is running in containers and turns on
-XX:+PortableSharedCache option by itself.
However, at the moment, OpenJ9 does not support cgroups v2 and cannot
detect whether it's running in containers on systems with cgroups v2.
This commit adds -XX:+PortableSharedCache to the JAVA_TOOL_OPTIONS
env var (which is read by OpenJ9) to specifically enable the portable
AOT feature.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>