Generate /etc/instanton.ld.so.cache
This cache can be used by downstream images that want to include
checkpointed processes by overwriting the default /etc/ld.so.cache
prior to starting a process that will be checkpointed.

This cache excludes any previously cached libraries that are located
under a directory named glibc-hwcaps. This directory contains
processor-specific libraries that, when loaded, make it unlikely that
a checkpointed process can be restored on an older (and probably
incompatible) processor.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>